Creating a conversion object
