+++ geoh5py/objects/__init__.py
@@ -20,7 +20,6 @@
 
 from .block_model import BlockModel
 from .cell_object import CellObject
-from .convert import Convert
 from .curve import Curve
 from .drape_model import DrapeModel
 from .drillhole import Drillhole
