+++ geoh5py/shared/conversion/base_conversion.py
@@ -1,0 +1,37 @@
+#  Copyright (c) 2022 Mira Geoscience Ltd.
+#
+#  This file is part of geoh5py.
+#
+#  geoh5py is free software: you can redistribute it and/or modify
+#  it under the terms of the GNU Lesser General Public License as published by
+#  the Free Software Foundation, either version 3 of the License, or
+#  (at your option) any later version.
+#
+#  geoh5py is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU Lesser General Public License for more details.
+#
+#  You should have received a copy of the GNU Lesser General Public License
+#  along with geoh5py.  If not, see <https://www.gnu.org/licenses/>.
+
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from ..entity import Entity
+
+
+class ConversionBase:  # pylint: disable=too-few-public-methods
+    def __init__(self, entity: Entity):
+        """
+        Converter class from an :obj:geoh5py.shared.entity.Entity to another.
+        :param entity: the entity to convert.
+        """
+        self._entity = entity
+
+    @property
+    def entity(self):
+        return self._entity
